+++ Bugsnag/Breadcrumbs/BSGNotificationBreadcrumbs.m
@@ -39,6 +39,7 @@
 #endif
         _breadcrumbSink = breadcrumbSink;
         _notificationNameMap = @{
+            @"NSProcessInfoThermalStateDidChangeNotification" : @"Thermal State Changed", // Using string to avoid availability issues
             NSUndoManagerDidRedoChangeNotification : @"Redo Operation",
             NSUndoManagerDidUndoChangeNotification : @"Undo Operation",
 #if TARGET_OS_TV
@@ -227,6 +228,20 @@
                                       object:nil];
         }
 #endif
+        
+#if TARGET_OS_IOS
+        [self.notificationCenter addObserver:self
+                                    selector:@selector(orientationDidChange:)
+                                        name:UIDeviceOrientationDidChangeNotification
+                                      object:nil];
+#endif
+        
+        if (@available(iOS 11.0, tvOS 11.0, *)) {
+            [self.notificationCenter addObserver:self
+                                        selector:@selector(thermalStateDidChange:)
+                                            name:NSProcessInfoThermalStateDidChangeNotification
+                                          object:nil];
+        }
     }
     
     // Navigation events
@@ -373,6 +388,49 @@ static NSString *nullStringIfBlank(NSString *str) {
     }
     [self addBreadcrumbWithType:BSGBreadcrumbTypeUser forNotificationName:notification.name metadata:metadata];
 #endif
+}
+
+#pragma mark -
+
+#if TARGET_OS_IOS
+
+- (void)orientationDidChange:(NSNotification *)notification {
+    UIDevice *device = notification.object;
+    
+    static UIDeviceOrientation previousOrientation;
+    if (device.orientation == UIDeviceOrientationUnknown ||
+        device.orientation == previousOrientation) {
+        return;
+    }
+    
+    NSMutableDictionary *metadata = [NSMutableDictionary dictionary];
+    metadata[@"from"] = BSGStringFromDeviceOrientation(previousOrientation);
+    metadata[@"to"] =  BSGStringFromDeviceOrientation(device.orientation);
+    previousOrientation = device.orientation;
+    
+    [self addBreadcrumbWithType:BSGBreadcrumbTypeState
+            forNotificationName:notification.name
+                       metadata:metadata];
+}
+
+#endif
+
+- (void)thermalStateDidChange:(NSNotification *)notification API_AVAILABLE(ios(11.0), tvos(11.0)) {
+    NSProcessInfo *processInfo = notification.object;
+    
+    static NSProcessInfoThermalState previousThermalState;
+    if (processInfo.thermalState == previousThermalState) {
+        return;
+    }
+    
+    NSMutableDictionary *metadata = [NSMutableDictionary dictionary];
+    metadata[@"from"] = BSGStringFromThermalState(previousThermalState);
+    metadata[@"to"] = BSGStringFromThermalState(processInfo.thermalState);
+    previousThermalState = processInfo.thermalState;
+    
+    [self addBreadcrumbWithType:BSGBreadcrumbTypeState
+            forNotificationName:notification.name
+                       metadata:metadata];
 }
 
 @end