+++ Tests/BugsnagTests/BSGOutOfMemoryTests.m
@@ -1,12 +1,12 @@
 #import <XCTest/XCTest.h>
 
 #import "BSGFileLocations.h"
+#import "BSGRunContext.h"
 #import "BSG_KSCrashState.h"
 #import "BSG_KSSystemInfo.h"
 #import "Bugsnag.h"
 #import "BugsnagClient+Private.h"
 #import "BugsnagConfiguration.h"
-#import "BugsnagKVStoreObjC.h"
 #import "BugsnagSystemState.h"
 #import "BugsnagTestConstants.h"
 
@@ -43,7 +43,6 @@
     NSDictionary *app = [state objectForKey:@"app"];
     XCTAssertNotNil([app objectForKey:@"bundleVersion"]);
     XCTAssertNotNil([app objectForKey:@"id"]);
-    XCTAssertNotNil([app objectForKey:@"inForeground"]);
     XCTAssertNotNil([app objectForKey:@"version"]);
     XCTAssertNotNil([app objectForKey:@"name"]);
     XCTAssertEqualObjects([app valueForKey:@"codeBundleId"], @"codeBundleIdHere");
@@ -71,59 +70,66 @@
 }
 
 - (void)testLastLaunchTerminatedUnexpectedly {
-    BugsnagKVStore *kvStore = [BugsnagKVStore new];
-    BugsnagSystemState *(^ systemState)() = ^{
-        return [[BugsnagSystemState alloc] initWithConfiguration:[[BugsnagConfiguration alloc] initWithApiKey:DUMMY_APIKEY_32CHAR_1]];
-    };
+    if (!bsg_runContext) {
+        BSGRunContextInit(BSGFileLocations.current.runContext.fileSystemRepresentation);
+    }
+    const struct BSGRunContext *oldContext = bsg_lastRunContext;
+    struct BSGRunContext lastRunContext = *bsg_runContext;
+    bsg_lastRunContext = &lastRunContext;
 
     // Debugger active
     
-    [kvStore setBoolean:true forKey:SYSTEMSTATE_APP_DEBUGGER_IS_ACTIVE];
-    [kvStore setBoolean:true forKey:SYSTEMSTATE_APP_WAS_TERMINATED];
-    [kvStore setBoolean:true forKey:SYSTEMSTATE_APP_IS_IN_FOREGROUND];
-    XCTAssertFalse(systemState().lastLaunchTerminatedUnexpectedly);
+    lastRunContext.isDebuggerAttached = true;
+    lastRunContext.isTerminating = true;
+    lastRunContext.isForeground = true;
+    XCTAssertFalse(BSGRunContextWasKilled());
 
-    [kvStore setBoolean:true forKey:SYSTEMSTATE_APP_DEBUGGER_IS_ACTIVE];
-    [kvStore setBoolean:true forKey:SYSTEMSTATE_APP_WAS_TERMINATED];
-    [kvStore setBoolean:false forKey:SYSTEMSTATE_APP_IS_IN_FOREGROUND];
-    XCTAssertFalse(systemState().lastLaunchTerminatedUnexpectedly);
+    lastRunContext.isDebuggerAttached = true;
+    lastRunContext.isTerminating = true;
+    lastRunContext.isForeground = false;
+    XCTAssertFalse(BSGRunContextWasKilled());
 
-    [kvStore setBoolean:true forKey:SYSTEMSTATE_APP_DEBUGGER_IS_ACTIVE];
-    [kvStore setBoolean:false forKey:SYSTEMSTATE_APP_WAS_TERMINATED];
-    [kvStore setBoolean:true forKey:SYSTEMSTATE_APP_IS_IN_FOREGROUND];
-    XCTAssertFalse(systemState().lastLaunchTerminatedUnexpectedly);
+    lastRunContext.isDebuggerAttached = true;
+    lastRunContext.isTerminating = false;
+    lastRunContext.isForeground = true;
+    XCTAssertFalse(BSGRunContextWasKilled());
 
-    [kvStore setBoolean:true forKey:SYSTEMSTATE_APP_DEBUGGER_IS_ACTIVE];
-    [kvStore setBoolean:false forKey:SYSTEMSTATE_APP_WAS_TERMINATED];
-    [kvStore setBoolean:false forKey:SYSTEMSTATE_APP_IS_IN_FOREGROUND];
-    XCTAssertFalse(systemState().lastLaunchTerminatedUnexpectedly);
+    lastRunContext.isDebuggerAttached = true;
+    lastRunContext.isTerminating = false;
+    lastRunContext.isForeground = false;
+    XCTAssertFalse(BSGRunContextWasKilled());
 
     // Debugger inactive
 
-    [kvStore setBoolean:false forKey:SYSTEMSTATE_APP_DEBUGGER_IS_ACTIVE];
-    [kvStore setBoolean:true forKey:SYSTEMSTATE_APP_WAS_TERMINATED];
-    [kvStore setBoolean:true forKey:SYSTEMSTATE_APP_IS_IN_FOREGROUND];
-    XCTAssertFalse(systemState().lastLaunchTerminatedUnexpectedly);
+    lastRunContext.isDebuggerAttached = false;
+    lastRunContext.isTerminating = true;
+    lastRunContext.isForeground = true;
+    XCTAssertFalse(BSGRunContextWasKilled());
 
-    [kvStore setBoolean:false forKey:SYSTEMSTATE_APP_DEBUGGER_IS_ACTIVE];
-    [kvStore setBoolean:true forKey:SYSTEMSTATE_APP_WAS_TERMINATED];
-    [kvStore setBoolean:false forKey:SYSTEMSTATE_APP_IS_IN_FOREGROUND];
-    XCTAssertFalse(systemState().lastLaunchTerminatedUnexpectedly);
+    lastRunContext.isDebuggerAttached = false;
+    lastRunContext.isTerminating = true;
+    lastRunContext.isForeground = false;
+    XCTAssertFalse(BSGRunContextWasKilled());
 
-    [kvStore setBoolean:false forKey:SYSTEMSTATE_APP_DEBUGGER_IS_ACTIVE];
-    [kvStore setBoolean:false forKey:SYSTEMSTATE_APP_WAS_TERMINATED];
-    [kvStore setBoolean:true forKey:SYSTEMSTATE_APP_IS_IN_FOREGROUND];
-    XCTAssertTrue(systemState().lastLaunchTerminatedUnexpectedly);
+    lastRunContext.isDebuggerAttached = false;
+    lastRunContext.isTerminating = false;
+    lastRunContext.isForeground = true;
+    XCTAssertTrue(BSGRunContextWasKilled());
+    
+    uuid_generate(lastRunContext.machoUUID);
+    XCTAssertFalse(BSGRunContextWasKilled());
+    uuid_copy(lastRunContext.machoUUID, bsg_runContext->machoUUID);
+    
+    lastRunContext.bootTime = 0;
+    XCTAssertFalse(BSGRunContextWasKilled());
+    lastRunContext.bootTime = bsg_runContext->bootTime;
 
-    [kvStore setBoolean:false forKey:SYSTEMSTATE_APP_DEBUGGER_IS_ACTIVE];
-    [kvStore setBoolean:false forKey:SYSTEMSTATE_APP_WAS_TERMINATED];
-    [kvStore setBoolean:false forKey:SYSTEMSTATE_APP_IS_IN_FOREGROUND];
-    XCTAssertFalse(systemState().lastLaunchTerminatedUnexpectedly);
-
-    // Delete keys so they don't interfere with other tests.
-    [kvStore deleteKey:SYSTEMSTATE_APP_DEBUGGER_IS_ACTIVE];
-    [kvStore deleteKey:SYSTEMSTATE_APP_WAS_TERMINATED];
-    [kvStore deleteKey:SYSTEMSTATE_APP_IS_IN_FOREGROUND];
+    lastRunContext.isDebuggerAttached = false;
+    lastRunContext.isTerminating = false;
+    lastRunContext.isForeground = false;
+    XCTAssertFalse(BSGRunContextWasKilled());
+    
+    bsg_lastRunContext = oldContext;
 }
 
 @end