+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+## 6.17.0 (2022-05-11)
+
+### Enhancements
+
+* Add `device.freeMemory` to OOM and Thermal Kill events on iOS 13 and later.
+  [#1357](https://github.com/bugsnag/bugsnag-cocoa/pull/1357)
+
+* Add `device.time` to OOM and Thermal Kill events.
+  [#1355](https://github.com/bugsnag/bugsnag-cocoa/pull/1355)
+
 ## 6.16.8 (2022-05-04)
 
 ### Changes